backend/requirements.txt
Django>=5.0,<6.0
djangorestframework>=3.15,<4.0
djangorestframework-simplejwt>=5.3,<6.0
django-cors-headers>=4.3,<5.0
dj-database-url>=2.1,<3.0
python-dotenv>=1.0,<2.0
psycopg2-binary>=2.9,<3.0
google-auth>=2.0,<3.0
cryptography>=43.0.0,<44.0.0

